[Refactor] #42 DomainEvent를 단일 소스로 사용
