+++ common/src/main/java/com/ticketrush/global/config/KafkaConfig.java
@@ -1,6 +1,6 @@
 package com.ticketrush.global.config;
 
-import com.ticketrush.global.eventpublisher.topic.DomainEventEnvelope;
+import com.ticketrush.global.event.DomainEventEnvelope;
 import java.util.HashMap;
 import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
