+++ common/src/main/java/com/ticketrush/global/eventpublisher/KafkaDomainEventPublisher.java
@@ -1,9 +1,7 @@
 package com.ticketrush.global.eventpublisher;
 
 import com.ticketrush.global.event.DomainEvent;
-import com.ticketrush.global.eventpublisher.topic.DomainEventEnvelope;
-import com.ticketrush.global.eventpublisher.topic.KafkaPublishTarget;
-import com.ticketrush.global.eventpublisher.topic.KafkaResolver;
+import com.ticketrush.global.event.DomainEventEnvelope;
 import com.ticketrush.global.json.JsonConverter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -20,19 +18,21 @@ import org.springframework.stereotype.Component;
 public class KafkaDomainEventPublisher implements EventPublisher {
 
   private final KafkaTemplate<String, DomainEventEnvelope> kafkaTemplate;
-  private final KafkaResolver kafkaResolver;
   private final JsonConverter jsonConverter;
 
   @Override
   public void publish(DomainEvent event) {
-    KafkaPublishTarget target = kafkaResolver.resolve(event);
+    // 1. Payload 직렬화
     String payload = jsonConverter.serialize(event);
-    DomainEventEnvelope envelope = DomainEventEnvelope.of(event, target.topic(), payload);
 
+    // 2. Envelope 생성
+    DomainEventEnvelope envelope = DomainEventEnvelope.of(event, payload);
+
+    // 3. Message 빌드 시 event.topic()과 event.key()를 직접 사용
     var message =
         MessageBuilder.withPayload(envelope)
-            .setHeader(KafkaHeaders.TOPIC, target.topic())
-            .setHeader(KafkaHeaders.KEY, target.kafkaKey())
+            .setHeader(KafkaHeaders.TOPIC, event.topic())
+            .setHeader(KafkaHeaders.KEY, event.key())
             .setHeader("eventType", envelope.eventType())
             .setHeader("eventId", envelope.eventId())
             .build();
@@ -44,7 +44,7 @@ public class KafkaDomainEventPublisher implements EventPublisher {
               if (ex != null) {
                 log.error(
                     "Failed to publish event to Kafka topic: {}, eventType={}, eventId={}",
-                    target.topic(),
+                    event.topic(),
                     envelope.eventType(),
                     envelope.eventId(),
                     ex);
